修复最大内存限制数字可能溢出的问题
将int改为了long

+++ runner.h
@@ -36,7 +36,7 @@ struct result {
 
 struct config {
     int max_cpu_time;
-    int max_memory;
+    long max_memory;
     char *path;
     char *in_file;
     char *out_file;
